requirements.txt
backcall==0.2.0
certifi==2021.10.8
charset-normalizer==2.0.11
decorator==5.1.1
idna==3.3
ipdb==0.13.9
ipython==7.31.1
jedi==0.18.1
matplotlib-inline==0.1.3
numpy==1.21.5
pandas==1.3.5
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.26
ptyprocess==0.7.0
Pygments==2.11.2
PyMySQL==1.0.2
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.1
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
six==1.16.0
toml==0.10.2
traitlets==5.1.1
urllib3==1.26.8
wcwidth==0.2.5
